improve qgb_network and move it to separate folder
