+++ e2e/qgb/network/qgb_network.go
@@ -1,4 +1,4 @@
-package e2e
+package network
 
 import (
 	"context"
