+++ e2e/qgb/network/test_commons.go
@@ -1,4 +1,4 @@
-package e2e
+package network
 
 import (
 	"context"
